refactoring user-section
