+++ src/user-content/user-page/user-index/user-actionline-item/action-line.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import styled from 'styled-components'
 
-import Button from '../../../ui-library/button-click/button'
+import Button from '../../../../ui-library/button-click/button'
 
 interface ActionitemProps {
   header?: string
