+++ src/user-content/user-page/user-index/user-header-block/user-header.tsx
@@ -2,11 +2,11 @@ import { userInfo } from 'os'
 import React from 'react'
 import styled from 'styled-components'
 
-import cover from '../../../assets/pages/chess-play/cover.jpg'
-import Settings from '../../../assets/user/settings.png'
-import Button from '../../../ui-library/button-pictures/button-pictures'
-import Logo from '../../../ui-library/logo/logo'
-import Upload from '../../../ui-library/upload-file/upload-file'
+import cover from '../../../../assets/pages/chess-play/cover.jpg'
+import Settings from '../../../../assets/user/settings.png'
+import Button from '../../../../ui-library/button-pictures/button-pictures'
+import Logo from '../../../../ui-library/logo/logo'
+import Upload from '../../../../ui-library/upload-file/upload-file'
 function onUploadCover(e: any, props: any) {
   if (e.target.files.length) {
     alert('upload')
