+++ src/user-content/user-page/user-index/user-people-block/user-people-block.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import styled from 'styled-components'
 
-import H3 from '../../../ui-library/h3/h3'
+import H3 from '../../../../ui-library/h3/h3'
 
 interface UserPeopleBlockProps {
   spanlength: number | undefined
